+++ design-system/motion/examples-utils/containers.tsx
@@ -1,24 +1,30 @@
 /** @jsx jsx */
 import React, { useState } from 'react';
 
-import { jsx } from '@emotion/react';
+import { css, jsx } from '@emotion/react';
 
 import Button from '@atlaskit/button/custom-theme-button';
+
+const containerStyles = css({
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+});
 
 export const Centered = ({
   as: As = 'div',
   ...props
-}: React.HTMLProps<HTMLDivElement> & { as?: string }) => (
-  <As
-    css={{
-      display: 'flex',
-      alignItems: 'center',
-      justifyContent: 'center',
-    }}
-    // eslint-disable-next-line @repo/internal/react/no-unsafe-spread-props
-    {...props}
-  />
-);
+}: React.HTMLProps<HTMLDivElement> & { as?: keyof JSX.IntrinsicElements }) => {
+  const Component = As as React.ElementType;
+
+  return (
+    <Component
+      css={containerStyles}
+      // eslint-disable-next-line @repo/internal/react/no-unsafe-spread-props
+      {...props}
+    />
+  );
+};
 
 export const RetryContainer = (props: { children: React.ReactNode }) => {
   const [count, setCount] = useState(0);