+++ design-system/motion/examples/constellation/motion-slide-in.tsx
@@ -1,0 +1,69 @@
+/** @jsx jsx */
+import { useState } from 'react';
+
+import { css, jsx } from '@emotion/react';
+
+import ButtonGroup from '@atlaskit/button/button-group';
+import Button from '@atlaskit/button/standard-button';
+import { type Direction, ExitingPersistence, SlideIn } from '@atlaskit/motion';
+import { token } from '@atlaskit/tokens';
+
+import { Block, Centered, RetryContainer } from '../../examples-utils';
+import { Fade } from '../../src/entering/types';
+
+const MotionSlideInExample = () => {
+  const [isIn, setIsIn] = useState(true);
+  const [fromIndex, setFromIndex] = useState(0);
+  const [fadeIndex, setFadeIndex] = useState(0);
+
+  return (
+    <RetryContainer>
+      <div css={containerStyles}>
+        <ButtonGroup>
+          <Button onClick={() => setIsIn((prev) => !prev)}>
+            {isIn ? 'Exit' : 'Enter'}
+          </Button>
+          <Button
+            onClick={() => setFromIndex((prev) => (prev + 1) % forms.length)}
+          >
+            From {forms[fromIndex]}
+          </Button>
+          <Button
+            onClick={() => setFadeIndex((prev) => (prev + 1) % fades.length)}
+          >
+            Fade {fades[fadeIndex]}
+          </Button>
+        </ButtonGroup>
+
+        <Centered css={centeredStyles}>
+          <ExitingPersistence appear>
+            {isIn && (
+              <SlideIn enterFrom={forms[fromIndex]} fade={fades[fadeIndex]}>
+                {(props) => <Block {...props} css={blockStyles} />}
+              </SlideIn>
+            )}
+          </ExitingPersistence>
+        </Centered>
+      </div>
+    </RetryContainer>
+  );
+};
+
+const forms: Direction[] = ['top', 'right', 'bottom', 'left'];
+const fades: Fade[] = ['none', 'in', 'out', 'inout'];
+
+const containerStyles = css({ textAlign: 'center' });
+
+const centeredStyles = css({
+  height: '300px',
+  margin: `${token('space.0', '0')} auto`,
+  position: 'relative',
+  overflow: 'hidden',
+});
+
+const blockStyles = css({
+  width: '95%',
+  height: '95%',
+});
+
+export default MotionSlideInExample;