+++ design-system/motion/src/entering/staggered-entrance.tsx
@@ -14,14 +14,15 @@ export interface StaggeredEntranceProps {
    */
   delayStep?: number;
 
+  // eslint-disable-next-line jsdoc/require-asterisk-prefix, jsdoc/check-alignment
   /**
-   * Number of columns the children elements will be displayed over.
-   * Use `"responsive"` to have it calculate dynamically on the client side.
-   *
-   * __NOTE:__ This has a big caveat that the elements will be invisible until the client side Javascript executes.
-   * If you have a fixed grid or list, set this to a specific number.
-   * Defaults to `"responsive"`.
-   */
+   Number of columns the children elements will be displayed over.
+   Use `"responsive"` to have it calculate dynamically on the client side.
+
+   __NOTE:__ This has a big caveat that the elements will be invisible until the client side Javascript executes.
+   If you have a fixed grid or list, set this to a specific number.
+   Defaults to `"responsive"`.
+  */
   columns?: number | 'responsive';
 
   /**