+++ editor/editor-core/examples/5-full-page-with-link-picker.tsx
@@ -1,30 +1,18 @@
 import React, { useMemo } from 'react';
 
-import { createSearchProvider, Scope } from '@atlassian/search-provider';
-import { AtlassianLinkPickerPlugin } from '@atlassian/link-picker-atlassian-plugin';
 import { useLinkPickerEditorProps } from '@atlassian/link-picker-plugins/editor';
 import { SmartCardProvider, CardClient } from '@atlaskit/link-provider';
 
 import { default as FullPageExample } from './5-full-page';
 
-const searchProvider = Promise.resolve(
-  createSearchProvider(
-    'DUMMY-a5a01d21-1cc3-4f29-9565-f2bb8cd969f5',
-    Scope.ConfluencePageBlog,
-    'https://pug.jira-dev.com/gateway/api/xpsearch-aggregator',
-  ),
-);
-
-const atlassianPlugin = new AtlassianLinkPickerPlugin({
-  searchProvider,
-  activityClientEndpoint: 'https://pug.jira-dev.com/gateway/api/graphql',
-});
-
 const smartCardClient = new CardClient('staging');
 
 const FullPageWithLinkPicker = () => {
-  const linkPicker = useLinkPickerEditorProps({
-    plugins: [atlassianPlugin],
+  const linkPicker = useLinkPickerEditorProps(undefined, {
+    product: 'confluence',
+    activityClientEndpoint: 'https://pug.jira-dev.com/gateway/api/graphql',
+    cloudId: 'a957adff-45b0-4f4f-8669-b640ed9973b6',
+    aggregatorUrl: 'https://pug.jira-dev.com/gateway/api/xpsearch-aggregator',
   });
 
   return (
@@ -43,11 +31,8 @@ const FullPageWithLinkPicker = () => {
 
 export default () => {
   const featureFlags = useMemo(() => {
-    const flagKey = 'disableLinkPickerPopupPositioningFix';
     return {
-      [flagKey]: Boolean(
-        new URLSearchParams(window.location.search).get(flagKey) ?? false,
-      ),
+      useLinkPickerAtlassianTabs: true,
     } as const;
   }, []);