+++ linking-platform/linking-common/src/ff.ts
@@ -3,13 +3,6 @@ export interface LinkingPlatformFeatureFlags {
   trackIframeDwellEvents?: boolean;
   useLinkPickerScrollingTabs?: boolean;
   /**
-   * Conditionally adds additional container styles to the link picker if the picker is going to appear outside the top boundary of the viewport.
-   * This feature will be enabled by default (not enabled by a feature flag) and we will disable the feature using the same feature flag that we will
-   * share with a long-term/proper fix in editor-core for their popup positioning.
-   * @atlaskit/link-picker-plugins/editor
-   */
-  disableLinkPickerPopupPositioningFix?: boolean;
-  /**
    * Enable an experiment on clickable State element on Hover Preview for Jira issue link.
    */
   enableActionableElement?: boolean;