+++ linking-platform/link-create/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/link-create
 
+## 0.0.3
+
+### Patch Changes
+
+- [`4b8910b3460`](https://bitbucket.org/atlassian/atlassian-frontend/commits/4b8910b3460) - Add link-create-confluence package
+
 ## 0.0.2
 
 ### Patch Changes